* Settings menu * Speed improvements to loading * Fixed bug related to opening and saving pack files where the header indicates additional info -> Folders were stripped from filenames incorrectly

+++ Common/PackFile.cs
@@ -312,7 +312,7 @@ namespace Common {
                 return result;
             }
         }
-        public long AdditionalInfo {
+        public UInt32 AdditionalInfo {
             get; set;
         }
